+++ src/TestSuite/TestCase.php
@@ -41,6 +41,7 @@ abstract class TestCase extends CakeTestCase
     /**
      * Called after every test method
      * @return void
+     * @throws \Throwable
      */
     public function tearDown(): void
     {