+++ tests/test_app/TestApp/webroot/js/test.js
@@ -3,7 +3,7 @@ function otherAlert()
     alert("Another alert");
 }
 
-$(function () {
-    var msg = "Ehi!";
+$(() => {
+    const msg = "Ehi!";
     alert(msg);
 });